requirements.txt
streamlit>=1.28.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
biopython>=1.81
tensorflow>=2.13.0
protobuf>=3.20.0,<5.0.0
